requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# ── Audio / CV ────────────────────────────────────────────────────────────────
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.26.0
Pillow>=10.2.0
opencv-python>=4.9.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
timm>=0.9.16

# ── NLP ───────────────────────────────────────────────────────────────────────
openai-whisper>=20231117
transformers>=4.38.0
sentence-transformers>=2.5.0
spacy>=3.7.0
nltk>=3.8.1

# ── Scraping ──────────────────────────────────────────────────────────────────
yt-dlp>=2024.3.10
telethon>=1.34.0
feedparser>=6.0.11
playwright>=1.42.0
requests>=2.31.0
beautifulsoup4>=4.12.3

# ── UI & API ──────────────────────────────────────────────────────────────────
gradio>=4.20.0
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9

# ── Reports ───────────────────────────────────────────────────────────────────
reportlab>=4.1.0
networkx>=3.2.1
pyvis>=0.3.2

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.2
pandas>=2.2.0
python-dotenv>=1.0.1
loguru>=0.7.2
pytest>=8.0.0
